+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Release 0.10.0 (unreleased)
 * Fix too strict overlapping path check (#684)
 * Show complete URL of child manifests (#683)
 * Show remote name when using default remote (#445)
+* Select HEAD branch as default in git (#689)
 
 Release 0.9.1 (released 2024-12-31)
 ===================================